feat: alert위치를 버튼 컨트롤러로 변경 및 locationDelegate 타입 새로운 파일로 위치변경
